refactor: remove user directory for entities
