+++ Frontend/src/entities/api/getAllBlogs.ts
@@ -1,5 +1,5 @@
-import { api } from "../../../shared/api/axios";
-import type { Blog } from "../model/blog";
+import {api} from "../../shared/api/axios.ts";
+import type {Blog} from "../model/blog.ts";
 
 export async function getAllBlogs() {
     const response = await api.get<Blog[]>("/blog/");
