+++ Frontend/src/entities/api/getBlog.ts
@@ -1,6 +1,6 @@
-import type { Guid } from "guid-typescript";
-import { api } from "../../../shared/api/axios";
-import type { Blog } from "../model/blog";
+import type {Guid} from "guid-typescript";
+import {api} from "../../shared/api/axios.ts";
+import type {Blog} from "../model/blog.ts";
 
 export async function getBlog(blogGuid: Guid) {
     const response = await api.get<Blog>(`/blog/${blogGuid}`);
