+++ Frontend/src/entities/model/blog.ts
@@ -1,5 +1,5 @@
 import type {Guid} from "guid-typescript";
-import type {User} from "./user";
+import type {User} from "./user.ts";
 
 export type Blog = {
     id: Guid;
